Connecting ML with app, dark/light mode, using resources for text, PL and ENG language, removing unnecessary imports, usersettings in viemodel implementation

+++ app/src/main/java/com/cnn/mushroom/data/MushroomEntity.kt
@@ -8,7 +8,7 @@ data class MushroomEntity(
 
     val imagePath: String,
     val name: String,
-    val timestamp: Int,
+    val timestamp: Long,
     val confidenceScore: Float? = null,
     val isEdible: Boolean? = null
 ) {
